+++ clang-tools-extra/test/clang-tidy/checkers/modernize/use-nullptr-c23.m
@@ -1,0 +1,7 @@
+// RUN: %check_clang_tidy %s modernize-use-nullptr %t -- -- -std=c23
+
+#define NULL 0
+
+void test_assignment() {
+  int *p1 = NULL;
+}